+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/repository/domain/message/internal/ReplyMessageEntity.kt
@@ -106,6 +106,8 @@ internal data class ReplyMessageInnerEntity(
     val restrictedVisibility: List<String> = emptyList(),
     /** Info about the reminder for the message **/
     val reminder: ReminderInfoEntity? = null,
+    /** The role of the member(who sent the message) in the channel */
+    val channelRole: String? = null,
 )
 
 internal const val REPLY_MESSAGE_ENTITY_TABLE_NAME = "stream_chat_reply_message"